+++ src/accounts/utils/signMessage.ts
@@ -1,7 +1,6 @@
-import { sign, Signature } from '@noble/secp256k1'
-
 import type { Hex } from '../../types'
 import { hashMessage } from '../../utils'
+import { sign } from './sign'
 
 export type SignMessageParameters = {
   message: string
@@ -13,7 +12,5 @@ export async function signMessage({
   message,
   privateKey,
 }: SignMessageParameters): Promise<SignMessageReturnType> {
-  const messageHash = hashMessage(message)
-  const signature = await sign(messageHash.slice(2), privateKey.slice(2))
-  return `0x${Signature.fromHex(signature).toCompactHex()}`
+  return sign({ hash: hashMessage(message), privateKey })
 }
